Fix for the way popup menus were handled in elemental app option to hide value indicators on range sliders

+++ GameCommon/UIX/UIXScrollableSection.cpp
@@ -114,7 +114,7 @@ void	UIXScrollbar::OnMouseWheel(float fOffset)
 	
 	mScrollPosition += (int)(fOffset);
 
-	mScrollPositionScreen = mScrollPosition * mfHeightPerUnit;
+	mScrollPositionScreen = (int)( mScrollPosition * mfHeightPerUnit );
 	if (mScrollPositionScreen < 0)
 	{
 		mScrollPositionScreen = 0;
